+++ index.html
@@ -58,106 +58,23 @@
         <a class="primary-button" href="#Contact-form"> Start collaboration</a>
         <div class="footer"></div>
       </div>
+      <div class="area" >
+        <ul class="circles">
+                <li></li>
+                <li></li>
+                <li></li>
+                <li></li>
+                <li></li>
+                <li></li>
+                <li></li>
+                <li></li>
+                <li></li>
+                <li></li>
+        </ul>
+  </div >
     </section>
-    <li></li>
     <section id="portfolio" class="work">
       <h2 class="work-heading">Projects</h2>
-      <div class="project p1">
-        <img src="images/project1.png" alt="First Project image" />
-        <h3>Aebinar</h3>
-        <ul class="list-tech">
-          <li>HTML</li>
-          <li>CSS</li>
-          <li>JavaScript</li>
-        </ul>
-        <button class="projectBtn" type="button">
-          See this project <i class="fa fa-arrow-right"></i>
-        </button>
-      </div>
-      <div class="project p2">
-        <img src="images/project2.png" alt="Second Project image" />
-        <h3>Air Quality Index</h3>
-        <ul class="list-tech">
-          <li>CSS</li>
-          <li>React/Redux</li>
-          <li>RESTFUL APIs</li>
-          <li>Google APIs</li>
-        </ul>
-        <button class="projectBtn" type="button">
-          See this project<i class="fa fa-arrow-right"></i>
-        </button>
-      </div>
-      <div class="project p3">
-        <img src="images/project3.png" alt="Third Project image" />
-        <h3>Poke Play Card</h3>
-        <ul class="list-tech">
-          <li>HTML/CSS</li>
-          <li>JavaScript</li>
-          <li>RESFUL APIs</li>
-          <li>Webpack</li>
-        </ul>
-        <button class="projectBtn" type="button">
-          See this project<i class="fa fa-arrow-right"></i>
-        </button>
-      </div>
-      <div class="project p4">
-        <img src="images/project4.png" alt="Fourth Project Image" />
-        <h3>Book Store</h3>
-        <ul class="list-tech">
-          <li>CSS</li>
-          <li>React/Redux</li>
-          <li>RESTFUL APIs</li>
-        </ul>
-        <button class="projectBtn" type="button">
-          See this project<i class="fa fa-arrow-right"></i>
-        </button>
-      </div>
-      <div class="project p5">
-        <img src="images/project5.png" alt="Fifth Project Image" />
-        <h3>Space X hub</h3>
-        <ul class="list-tech">
-          <li>CSS</li>
-          <li>RESTFUL APIs</li>
-          <li>React/Redux</li>
-        </ul>
-        <button class="projectBtn" type="button">
-          See this project<i class="fa fa-arrow-right"></i>
-        </button>
-      </div>
-      <div class="project p6">
-        <img src="images/project6.png" alt="Sixth Project Image" />
-        <h3>Math Magician</h3>
-        <ul class="list-tech">
-          <li>CSS</li>
-          <li>React/Redux</li>
-        </ul>
-        <button class="projectBtn" type="button">
-          See this project<i class="fa fa-arrow-right"></i>
-        </button>
-      </div>
-      <div class="project p7">
-        <img src="images/project7.png" alt="Seventh Project Image" />
-        <h3>GStream</h3>
-        <ul class="list-tech">
-          <li>CSS</li>
-          <li>React</li>
-        </ul>
-        <button class="projectBtn" type="button">
-          See this project<i class="fa fa-arrow-right"></i>
-        </button>
-      </div>
-      <div class="project p8">
-        <img src="images/project8.png" alt="Eighth Project Image" />
-        <h3>Budge it!</h3>
-        <ul class="list-tech">
-          <li>Ruby</li>
-          <li>Ruby on Rails</li>
-          <li>Postgresql</li>
-        </ul>
-        <button class="projectBtn" type="button">
-          See this project<i class="fa fa-arrow-right"></i>
-        </button>
-      </div>
     </section>
     <section id="about" class="about-us">
       <div class="about-us-content">
@@ -273,6 +190,6 @@
       </div>
     </footer>
     <script src="Js/script.js"></script>
-    <script src="Js/model.js"></script>
+    <script type="module" src="Js/projects.js" defer></script>
   </body>
 </html>
