+++ style.css
@@ -202,7 +202,6 @@ ul {
   font-size: 50px;
   margin-bottom: 20px;
   font-weight: 400;
-  line-height: 50px;
   letter-spacing: 0;
   text-align: left;
 }
@@ -210,9 +209,13 @@ ul {
 .work .project {
   display: flex;
   flex-direction: column;
-  background: #1a2236;
+  background: transparent;
   justify-content: flex-start;
   color: #fff;
+}
+
+.work .projectBtns {
+  background-color: #1a2236;
 }
 
 .project img {
@@ -233,7 +236,9 @@ ul {
 .project .list-tech {
   list-style: none;
   margin: 20px 20px 0 20px;
-  display: inline;
+  display: flex;
+  gap: 6px;
+  flex-wrap: wrap;
 }
 
 /* popup-start */
@@ -242,7 +247,6 @@ ul {
   border: 1px solid #1a2236;
   padding: 10px 4%;
   margin-top: 5px;
-  margin-right: 10px;
   background-color: #fff;
   font-family: 'Poppins', sans-serif;
   font-style: normal;
@@ -535,7 +539,8 @@ ul.pop-tech-list {
   list-style: none;
   display: flex;
   flex-wrap: wrap;
-  margin: 5px 0;
+  gap: 6px;
+  padding: 5px;
 }
 
 img.pop-img {
@@ -725,6 +730,123 @@ span.alert.active {
     display: none;
   }
 
+
+.circles{
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  overflow: hidden;
+}
+
+.circles li{
+  position: absolute;
+  display: block;
+  list-style: none;
+  width: 20px;
+  height: 20px;
+  background: #e2505c87;
+  animation: animateCircles 25s linear infinite;
+  bottom: -150px;
+}
+
+.circles li:nth-child(1){
+  left: 25%;
+  width: 80px;
+  height: 80px;
+  background: #febf528f;
+  animation-delay: 0s;
+}
+
+
+.circles li:nth-child(2){
+  left: 10%;
+  width: 20px;
+  height: 20px;
+  animation-delay: 2s;
+  animation-duration: 12s;
+}
+
+.circles li:nth-child(3){
+  left: 70%;
+  width: 20px;
+  height: 20px;
+  animation-delay: 4s;
+}
+
+.circles li:nth-child(4){
+  left: 40%;
+  width: 60px;
+  height: 60px;
+  background: #febf528f;
+  animation-delay: 0s;
+  animation-duration: 18s;
+}
+
+.circles li:nth-child(5){
+  left: 65%;
+  width: 20px;
+  height: 20px;
+  animation-delay: 0s;
+}
+
+.circles li:nth-child(6){
+  left: 75%;
+  width: 110px;
+  height: 110px;
+  animation-delay: 3s;
+}
+
+.circles li:nth-child(7){
+  left: 35%;
+  width: 150px;
+  height: 150px;
+  background: #1c518a75;
+  animation-delay: 7s;
+}
+
+.circles li:nth-child(8){
+  left: 50%;
+  width: 25px;
+  height: 25px;
+  animation-delay: 15s;
+  background: #febf528f;
+  animation-duration: 45s;
+}
+
+.circles li:nth-child(9){
+  left: 20%;
+  width: 15px;
+  height: 15px;
+  animation-delay: 2s;
+  animation-duration: 35s;
+}
+
+.circles li:nth-child(10){
+  left: 85%;
+  width: 150px;
+  height: 150px;
+  animation-delay: 0s;
+  animation-duration: 11s;
+}
+
+@keyframes animateCircles {
+
+  0%{
+      transform: translateY(0) rotate(0deg);
+      opacity: 1;
+      border-radius: 0;
+  }
+
+  100%{
+      transform: translateY(-1000px) rotate(720deg);
+      opacity: 0;
+      border-radius: 50%;
+  }
+
+}
+
   /* work desktop version */
   .work {
     margin: 0;
@@ -747,16 +869,38 @@ span.alert.active {
     background-repeat: no-repeat;
   }
 
+  .project {
+    position: relative;
+    transition: all 0.5s ease-out;    
+  }
+
   .work h2.work-heading {
     grid-column: 1/3;
     grid-row: 1/4;
-    padding: 0;
-    margin: 0;
-    font-style: normal;
     font-size: 110px;
-    color: #1a2236;
+    font-weight: bolder;
     transform: rotate(-90deg);
+    color: rgba(225,225,225, .01);
+  background-image: url("https://images.unsplash.com/photo-1499195333224-3ce974eecb47?ixlib=rb-0.3.5&ixid=eyJhcHBfaWQiOjEyMDd9&s=2cf549433129d4227d1879347b9e78ce&auto=format&fit=crop&w=1248&q=80");
+  background-repeat: repeat;
+  -webkit-background-clip:text;
+  animation: animateHeading 15s ease-in-out infinite;
   }
+
+  @keyframes animateHeading {
+    0%, 100% {
+      background-position: left top;
+    }
+    25%{
+      background-position: right bottom;
+     }
+    50% {
+      background-position: left bottom;
+    }
+    75% {
+      background-position: right top;
+    }   
+}
 
   .work .p1 {
     width: 100%;
@@ -775,6 +919,7 @@ span.alert.active {
     margin-left: 25%;
     grid-column: 5/8;
     grid-row: 4/7;
+    z-index: 7;
   }
 
   .work .p4 {
@@ -785,7 +930,7 @@ span.alert.active {
   }
 
   .p4 img {
-    height: 100%;
+    height: auto;
   }
 
   .work .p5 {
@@ -794,39 +939,37 @@ span.alert.active {
     grid-row: 8/9;
   }
 
-  .p2 h3,
-  .p4 h3,
-  .p5 h3,
-  .p6 h3,
-  .p7 h3,
-  .p8 h3 {
-    display: none;
+  .work .p6 {
+    width: 100%;
+    grid-column: 3/5;
+    grid-row: 8/9;
   }
 
-  .p2 .list-tech,
-  .p4 .list-tech,
-  .p5 .list-tech,
-  .p6 .list-tech,
-  .p7 .list-tech,
-  .p8 .list-tech {
-    display: none;
+  .work .p7 {
+    width: 70%;
+    grid-column: 5/7;
+    grid-row: 3/5;
+    transform: translateY(-1rem);
+    z-index: 3;
   }
 
-  .p2 button,
-  .p4 button,
-  .p5 button,
-  .p6 button,
-  .p7 button,
-  .p8 button {
-    display: none;
+  .work .p8 {
+    width: 100%;
+    grid-column: 6/8;
+    grid-row: 7/7;
   }
 
-  .p3 h3 {
+  .work .p9 {
+    width: 100%;
+    grid-column-end: span 2;
+  }
+
+  .projectBtns h3 {
     padding: 0;
     font-size: 14px;
   }
 
-  .p3 .list-tech {
+  .projectBtns .list-tech {
     padding: 0;
     margin-top: 3px;
     font-size: 10px;
@@ -836,12 +979,12 @@ span.alert.active {
     font-size: 16px;
   }
 
-  .p3 .list-tech li {
+  .projectBtns .list-tech li {
     padding: 5px;
     margin-right: 3px;
   }
 
-  .p3 button {
+  .projectBtns button {
     margin-top: 6px;
     font-size: 11px;
     padding: 1px 7px;
@@ -852,9 +995,29 @@ span.alert.active {
     font-size: 8px;
   }
 
-  .p6 img,
-  .p8 img {
-    display: none;
+  .projectBtns {
+    background: #1a2236;
+  }
+
+  .project:hover {
+    transform: translate(1rem,2rem) scale(1.1);
+    z-index: 10;
+  }
+  
+  .project:nth-child(n+3) .projectBtns {
+    opacity: 0;
+    height: 0;
+    position: absolute;
+    border: #1a2236;
+    background: #1a2236a9;
+  }
+
+  .project:hover .projectBtns {
+    opacity: 1;
+    height: auto;
+    width: 100%;
+    bottom: 0;
+    transition: all 0.5s ease-out;    
   }
 
   /* about-us desktop version */
